backend/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
blinker==1.8.1
boto3==1.34.94
botocore==1.34.94
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
cssselect2==0.7.0
daphne==4.1.2
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
django-rest-passwordreset==1.4.1
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dotenv-python==0.0.1
drf-yasg==1.21.7
filters==1.3.2
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.7
lxml==5.2.1
MarkupSafe==2.1.5
msgpack==1.0.8
multidict==6.0.5
oscrypto==1.3.0
packaging==24.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.4
regex==2024.4.28
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.1
sendgrid==6.11.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
stripe==9.4.0
svglib==1.5.1
tinycss2==1.3.0
twilio==9.0.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
whitenoise==6.6.0
xhtml2pdf==0.2.15
yarl==1.9.4
zope.interface==6.3
